Create README.md
I'm not sure if I'm doing this right. I feel as though I have learned more but I'm still a bit lost. I am really trying tov balance it all and take the needed time to focus on this course. In this weeks homework, I understood how tov create the html code. I didv majority of this long hand. When it comes to combine all three language, I am not sure how to make it all function. I plan to continue on this homework after today to tweak it. I know it'll take time for this  to come effortlessly. In the meantime, I will deep dive into resources
that'll help vie improve.